+++ inkstitch.py
@@ -8,11 +8,16 @@ import sys
 from pathlib import Path  # to work with paths as objects
 from argparse import ArgumentParser  # to parse arguments and remove --extension
 
+if sys.version_info >= (3, 11):
+    import tomllib      # built-in in Python 3.11+
+else:
+    import tomli as tomllib
+
 import logging
 
 import lib.debug.utils as debug_utils
 import lib.debug.logging as debug_logging
-from lib.debug.utils import get_ini, safe_get    # mimic get method of dict with default value
+from lib.debug.utils import safe_get    # mimic get method of dict with default value
 
 # --------------------------------------------------------------------------------------------
 
@@ -20,7 +25,19 @@ SCRIPTDIR = Path(__file__).parent.absolute()
 
 logger = logging.getLogger("inkstitch")   # create module logger with name 'inkstitch'
 
-ini = get_ini()
+# TODO --- temporary --- catch old DEBUG.ini file and inform user to reformat it to DEBUG.toml
+old_debug_ini = SCRIPTDIR / "DEBUG.ini"
+if old_debug_ini.exists():
+    print("ERROR: old DEBUG.ini exists, please reformat it to DEBUG.toml and remove DEBUG.ini file", file=sys.stderr)
+    exit(1)
+# --- end of temporary ---
+
+debug_toml = SCRIPTDIR / "DEBUG.toml"
+if debug_toml.exists():
+    with debug_toml.open("rb") as f:
+        ini = tomllib.load(f)  # read DEBUG.toml file if exists, otherwise use default values in ini object
+else:
+    ini = {}
 # --------------------------------------------------------------------------------------------
 
 running_as_frozen = getattr(sys, 'frozen', None) is not None  # check if running from pyinstaller bundle
@@ -69,7 +86,7 @@ if not running_as_frozen:  # debugging/profiling only in development mode
     profiler_type = debug_utils.resolve_profiler_type(ini)  # read profile type from ini file or cmd line
 
     if running_from_inkscape:
-        # process creation of the Bash script - should be done before sys.path is modified, see below in prefere_pip_inkex
+        # process creation of the Bash script - should be done before sys.path is modified, see below in prefer_pip_inkex
         if safe_get(ini, "DEBUG", "create_bash_script", default=False):  # create script only if enabled in DEBUG.toml
             debug_utils.write_offline_debug_script(SCRIPTDIR, ini)
 
@@ -80,8 +97,8 @@ if not running_as_frozen:  # debugging/profiling only in development mode
 
     # prefer pip installed inkex over inkscape bundled inkex, pip version is bundled with Inkstitch
     # - must be be done before importing inkex
-    prefere_pip_inkex = safe_get(ini, "LIBRARY", "prefer_pip_inkex", default=True)
-    if prefere_pip_inkex and 'PYTHONPATH' in os.environ:
+    prefer_pip_inkex = safe_get(ini, "LIBRARY", "prefer_pip_inkex", default=True)
+    if prefer_pip_inkex and 'PYTHONPATH' in os.environ:
         debug_utils.reorder_sys_path()
 
 # enabling of debug depends on value of debug_type in DEBUG.toml file
