+++ lib/utils/paths.py
@@ -9,6 +9,13 @@ from os.path import dirname, realpath
 
 import appdirs
 
+from pathlib import Path
+
+if sys.version_info >= (3, 11):
+    import tomllib      # built-in in Python 3.11+
+else:
+    import tomli as tomllib
+
 
 def get_bundled_dir(name=None):
     if getattr(sys, 'frozen', None) is not None:
@@ -42,3 +49,13 @@ def get_user_dir(name=None):
         path = os.path.join(path, name)
 
     return path
+
+
+def get_ini():
+    debug_toml = Path(get_bundled_dir("DEBUG.toml"))
+    if debug_toml.exists():
+        with debug_toml.open("rb") as f:
+            ini = tomllib.load(f)  # read DEBUG.toml file if exists, otherwise use default values in ini object
+    else:
+        ini = {}
+    return ini
